+++ app/lib/shared/versions.dart
@@ -24,10 +24,10 @@ final RegExp runtimeVersionPattern = RegExp(r'^\d{4}\.\d{2}\.\d{2}$');
 /// when the version switch happens.
 const _acceptedRuntimeVersions = <String>[
   // The current [runtimeVersion].
-  '2025.03.10',
+  '2025.03.18',
   // Fallback runtime versions.
+  '2025.03.10',
   '2025.03.03',
-  '2025.02.25',
 ];
 
 /// Sets the current runtime versions.
@@ -62,8 +62,8 @@ bool shouldGCVersion(String version) =>
 
 // keep in-sync with SDK version in .mono_repo.yml and Dockerfile
 final String runtimeSdkVersion = '3.7.0';
-final String toolStableDartSdkVersion = '3.7.1';
-final String toolStableFlutterSdkVersion = '3.29.1';
+final String toolStableDartSdkVersion = '3.7.2';
+final String toolStableFlutterSdkVersion = '3.29.2';
 
 final semanticToolStableDartSdkVersion =
     Version.parse(toolStableDartSdkVersion);