+++ CHANGELOG.md
@@ -2,6 +2,9 @@ Important changes to data models, configuration, and migrations between each
 AppEngine version, listed here to ease deployment and troubleshooting.
 
 ## Next Release (replace with git tag when deployed)
+ * Bump runtimeVersion to `2025.03.18`.
+ * Upgraded stable Dart analysis SDK to `3.7.2`
+ * Upgraded stable Flutter analysis SDK to `3.29.2`.
 
 ## `20250311t101500-all`
  * Bump runtimeVersion to `2025.03.10`.